#10334: Migrate conj_bw to TTNN
