+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -633,7 +633,7 @@ Other Operations
 Backward Operations
 ===================
 
-.. autofunction:: tt_lib.tensor.conj_bw
+.. autofunction:: tt_lib.tensor.prod_bw
 
 .. autofunction:: tt_lib.tensor.complex_abs_bw
