+++ docs/source/ttnn/ttnn/ttnn/conj_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.conj_bw:
+
+ttnn.conj_bw
+###################
+
+.. autofunction:: ttnn.conj_bw
